Renamed github api related file
This was done because a name collision error started showing up with a github related dependency package in the recent version
